Update `semaBits` type to `uint16_t` for extended flag support and adjust related constants, enums, and methods accordingly.

+++ src/Parser/AstNodeId.h
@@ -2,7 +2,7 @@
 
 SWC_BEGIN_NAMESPACE()
 
-enum class AstNodeId : uint16_t
+enum class AstNodeId : uint8_t
 {
 #define SWC_NODE_DEF(__enum, __flags) __enum,
 #include "Parser/AstNodes.Def.inc"
